+++ sdk/ai/Azure.AI.VoiceLive/src/Customizations/VoiceProvider.cs
@@ -20,6 +20,20 @@ namespace Azure.AI.VoiceLive
 
         internal static VoiceProvider DeserializeVoiceProvider(JsonElement element, ModelReaderWriterOptions options)
         {
+            if (element.TryGetProperty("type"u8, out JsonElement discriminator))
+            {
+                switch (discriminator.GetString())
+                {
+                    case "openai":
+                        return OpenAIVoice.DeserializeOpenAIVoice(element, options);
+                    case "azure-custom":
+                    case "azure-standard":
+                    case "azure-personal":
+                        return AzureVoice.DeserializeAzureVoice(element, options);
+                    default:
+                        return null;
+                }
+            }
             return null;
         }
     }